Полировка кода для гитхаба
Комменты, рефакторинг, пример использования, мелочи.
